fix for bands>255

+++ tp1/demo-heights/shaders/fragment-shader.glsl
@@ -1,5 +1,4 @@
-        precision mediump float;
-
+precision highp float;
         varying vec2 vUv;
         varying vec3 vNormal;
         varying vec3 vWorldPosition;
